requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.0
praw>=7.8.0
sendgrid>=6.10.0
apscheduler>=3.10.0
requests>=2.31.0
anthropic>=0.39.0
